Implement core framework, report and test case
