+++ e2e/component/misc-utils/expectation-helper.ts
@@ -1,8 +1,8 @@
 import {PageHelper} from '../html/page-helper';
-import {browser, ElementFinder, ElementHelper} from 'protractor';
+import {ElementFinder} from 'protractor';
 import {StepLogger} from '../../../core/logger/step-logger';
 import {ValidationsHelper} from './validation-helper';
-import {HtmlHelper} from "./html-helper";
+import {HtmlHelper} from './html-helper';
 
 export class ExpectationHelper {
     /**
@@ -18,7 +18,7 @@ export class ExpectationHelper {
                                        toWait = true,
                                        timeout = PageHelper.DEFAULT_TIMEOUT) {
         stepLogger.subVerification(`${elementName} should display`);
-        await expect(await PageHelper.isElementDisplayed(targetElement, toWait, timeout))
+        expect(await PageHelper.isElementDisplayed(targetElement, toWait, timeout))
             .toBe(true,
                 ValidationsHelper.getDisplayedValidation(elementName));
     }
@@ -32,7 +32,7 @@ export class ExpectationHelper {
      */
     static async verifyHiddenStatus(targetElement: ElementFinder, elementName: string, stepLogger: StepLogger) {
         stepLogger.subVerification(`${elementName} should be hidden`);
-        await expect(await PageHelper.isElementHidden(targetElement))
+        expect(await PageHelper.isElementHidden(targetElement))
             .toBe(true,
                 ValidationsHelper.getDisplayedValidation(elementName));
     }
@@ -47,7 +47,7 @@ export class ExpectationHelper {
     static async verifyCheckboxIsChecked(targetElement: ElementFinder, elementName: string, stepLogger: StepLogger) {
         stepLogger.subVerification(`${elementName} should be checked`);
         const checkBoxStatus = await PageHelper.isElementSelected(targetElement);
-        await expect(checkBoxStatus).toBe(true, ValidationsHelper.getCheckedValidation(elementName));
+        expect(checkBoxStatus).toBe(true, ValidationsHelper.getCheckedValidation(elementName));
     }
 
     /**
@@ -59,7 +59,7 @@ export class ExpectationHelper {
      */
     static async verifyRemovedStatus(targetElement: ElementFinder, elementName: string, stepLogger: StepLogger) {
         stepLogger.subVerification(`${elementName} should be removed`);
-        await expect(await PageHelper.isElementHidden(targetElement))
+        expect(await PageHelper.isElementHidden(targetElement))
             .toBe(true,
                 ValidationsHelper.getDisplayedValidation(elementName));
     }
@@ -73,7 +73,7 @@ export class ExpectationHelper {
      */
     static async verifyEnabledStatus(targetElement: ElementFinder, elementName: string, stepLogger: StepLogger) {
         stepLogger.subVerification(`${elementName} should be enabled`);
-        await expect(await PageHelper.isElementEnabled(targetElement))
+        expect(await PageHelper.isElementEnabled(targetElement))
             .toBe(true,
                 ValidationsHelper.getEnabledValidation(elementName));
     }
@@ -87,7 +87,7 @@ export class ExpectationHelper {
      */
     static async verifyDisabledStatus(targetElement: ElementFinder, elementName: string, stepLogger: StepLogger) {
         stepLogger.subVerification(`${elementName} should be disabled`);
-        await expect(await PageHelper.isElementEnabled(targetElement))
+        expect(await PageHelper.isElementEnabled(targetElement))
             .toBe(false,
                 ValidationsHelper.getDisabledValidation(elementName));
     }
@@ -106,7 +106,7 @@ export class ExpectationHelper {
             targetElement,
             HtmlHelper.attributes.value,
         );
-        await expect(val === expectedValue)
+        expect(val === expectedValue)
             .toBe(true, ValidationsHelper.getFieldValueValidation(elementName, expectedValue));
     }
 
@@ -120,7 +120,7 @@ export class ExpectationHelper {
      */
     static async verifyText(targetElement: ElementFinder, elementName: string, expectedValue: string, stepLogger: StepLogger) {
         stepLogger.subVerification(`${elementName} should have exact text as ${expectedValue} `);
-        await expect((await PageHelper.getText(targetElement)).toLowerCase())
+        expect((await PageHelper.getText(targetElement)).toLowerCase())
             .toBe(expectedValue.toLowerCase(),
                 ValidationsHelper.getFieldShouldHaveValueValidation(elementName, expectedValue));
     }
@@ -135,7 +135,7 @@ export class ExpectationHelper {
      */
     static async verifyTextContains(targetElement: ElementFinder, elementName: string, expectedValue: string, stepLogger: StepLogger) {
         stepLogger.subVerification(`${elementName} should contain ${expectedValue} value`);
-        await expect((await PageHelper.getText(targetElement)).toLowerCase())
+        expect((await PageHelper.getText(targetElement)).toLowerCase())
             .toContain(expectedValue.toLowerCase(),
                 ValidationsHelper.getFieldShouldHaveValueValidation(elementName, expectedValue));
     }
@@ -149,7 +149,7 @@ export class ExpectationHelper {
      */
     static async verifyValueGreaterThan(actualValue: number, expectedValue: number, stepLogger: StepLogger) {
         stepLogger.subVerification(`${actualValue} should be grater than ${expectedValue} value`);
-        await expect(actualValue).toBeGreaterThan(
+        expect(actualValue).toBeGreaterThan(
             expectedValue, ValidationsHelper.getGraterThanValidation(actualValue, expectedValue));
     }
 
@@ -162,7 +162,7 @@ export class ExpectationHelper {
      */
     static async verifyValueLessOrEqualTo(actualValue: number, expectedValue: number, stepLogger: StepLogger) {
         stepLogger.subVerification(`${actualValue} should be less ot equal to ${expectedValue} value`);
-        await expect(actualValue).toBeLessThanOrEqual(
+        expect(actualValue).toBeLessThanOrEqual(
             expectedValue, ValidationsHelper.getLessThanOrEqualToValidation(actualValue, expectedValue));
     }
 
@@ -175,7 +175,7 @@ export class ExpectationHelper {
      */
     static async verifyValueEqualTo(actualValue: number, expectedValue: number, stepLogger: StepLogger) {
         stepLogger.subVerification(`${actualValue} should be equal to  ${expectedValue} value`);
-        await expect(actualValue).toEqual(
+        expect(actualValue).toEqual(
             expectedValue, ValidationsHelper.getEqualToValidation(actualValue, expectedValue));
     }
 
@@ -188,7 +188,7 @@ export class ExpectationHelper {
      */
     static async verifyValueNotEqualTo(actualValue: string, expectedValue: string, stepLogger: StepLogger) {
         stepLogger.subVerification(`${actualValue} should be not equal to  ${expectedValue} value`);
-        await expect(actualValue).not.toEqual(
+        expect(actualValue).not.toEqual(
             expectedValue, ValidationsHelper.getNotEqualToValidation(actualValue, expectedValue));
     }
 
@@ -202,7 +202,7 @@ export class ExpectationHelper {
     static async verifyCheckBoxNotSelected(targetElement: ElementFinder, elementName: string, stepLogger: StepLogger) {
         const actualValue = await targetElement.isSelected();
         stepLogger.subVerification(`${elementName} should not be selected`);
-        await expect(actualValue).toEqual(
+        expect(actualValue).toEqual(
             false, ValidationsHelper.getUnSelectedValidation(elementName));
     }
 
@@ -218,7 +218,7 @@ export class ExpectationHelper {
     static async verifyAttributeValue(targetElement: ElementFinder, attribute: string, expectedValue: string, stepLogger: StepLogger) {
         const actualValue = await PageHelper.getAttributeValue(targetElement, attribute);
         stepLogger.subVerification(`${actualValue} should be equal to  ${expectedValue} value`);
-        await expect(actualValue).toEqual(
+        expect(actualValue).toEqual(
             expectedValue, ValidationsHelper.getStringEqualToValidation(actualValue, expectedValue));
     }
 
@@ -231,7 +231,7 @@ export class ExpectationHelper {
      */
     static async verifyStringValueEqualTo(actualValue: string, expectedValue: string, stepLogger: StepLogger) {
         stepLogger.subVerification(`${actualValue} should be equal to  ${expectedValue} value`);
-        await expect(actualValue).toEqual(
+        expect(actualValue).toEqual(
             expectedValue, ValidationsHelper.getStringEqualToValidation(actualValue, expectedValue));
     }
 
@@ -244,7 +244,7 @@ export class ExpectationHelper {
      */
     static async verifyStringValueContain(actualValue: string, expectedValue: string, stepLogger: StepLogger) {
         stepLogger.subVerification(`${actualValue} should be contains to  ${expectedValue} value`);
-        await expect(actualValue).toContain(
+        expect(actualValue).toContain(
             expectedValue, ValidationsHelper.getStringEqualToValidation(actualValue, expectedValue));
     }
 
@@ -257,7 +257,7 @@ export class ExpectationHelper {
      */
     static async verifyStringValueNotContain(actualValue: string, expectedValue: string, stepLogger: StepLogger) {
         stepLogger.subVerification(`${actualValue} should be not contains to  ${expectedValue} value`);
-        await expect(actualValue).not.toContain(
+        expect(actualValue).not.toContain(
             expectedValue, ValidationsHelper.getStringEqualToValidation(actualValue, expectedValue));
     }
 
@@ -270,7 +270,7 @@ export class ExpectationHelper {
      */
     static async verifyActualValueContainsExpectedValue(actualValue: string, expectedValue: string, stepLogger: StepLogger) {
         stepLogger.subVerification(`${actualValue} should contain ${expectedValue} value`);
-        await expect(actualValue).toContain(expectedValue.toLowerCase(),
+        expect(actualValue).toContain(expectedValue.toLowerCase(),
             ValidationsHelper.getFieldShouldHaveValueValidation(actualValue, expectedValue));
     }
 
@@ -284,7 +284,7 @@ export class ExpectationHelper {
      */
     static async verifyContainsText(targetElement: ElementFinder, elementName: string, expectedValue: string, stepLogger: StepLogger) {
         stepLogger.subVerification(`${elementName} should have exact text as ${expectedValue} `);
-        await expect((await PageHelper.getText(targetElement)).toLowerCase())
+        expect((await PageHelper.getText(targetElement)).toLowerCase())
             .toContain(expectedValue.toLowerCase(),
                 ValidationsHelper.getFieldShouldHaveValueValidation(elementName, expectedValue));
     }
@@ -300,14 +300,14 @@ export class ExpectationHelper {
 
     static async verifyNotPresentStatus(targetElement: ElementFinder, elementName: string, stepLogger: StepLogger, toWait = true) {
         stepLogger.subVerification(`${elementName} should not be present`);
-        await expect(await PageHelper.isElementPresent(targetElement, toWait))
+        expect(await PageHelper.isElementPresent(targetElement, toWait))
             .toBe(false,
                 ValidationsHelper.getNotPresentValidation(elementName));
     }
 
     static async verifyNotDisplayedStatus(targetElement: ElementFinder, elementName: string, stepLogger: StepLogger) {
         stepLogger.subVerification(`${elementName} should not be displayed`);
-        await expect(await PageHelper.isElementDisplayed(targetElement, false))
+        expect(await PageHelper.isElementDisplayed(targetElement, false))
             .toBe(false,
                 ValidationsHelper.getDisplayedValidation(elementName));
     }
@@ -326,44 +326,15 @@ export class ExpectationHelper {
                                             toWait: boolean = true,
                                             timeout = PageHelper.DEFAULT_TIMEOUT) {
         stepLogger.subVerification(`${elementName} should present`);
-        await expect(await PageHelper.isElementPresent(targetElement, toWait, timeout))
+        expect(await PageHelper.isElementPresent(targetElement, toWait, timeout))
             .toBe(true,
                 ValidationsHelper.getPresentValidation(elementName));
-    }
-
-    /**
-     * Verify whether an element is displayed or not
-     * @param {ElementFinder} targetElement
-     * @param {string} elementName
-     * @param {StepLogger} stepLogger
-     * @param {boolean} toWait
-     * @returns {Promise<void>}
-     */
-    static async verifyNewWindowUrlContains(urlSubString: string, stepLogger: StepLogger) {
-        stepLogger.subVerification(`new window URl contains ${urlSubString}`);
-        await browser.getAllWindowHandles().then(async function (handles) {
-            await browser.switchTo().window(handles[1]).then(async function () {
-                await expect(browser.getCurrentUrl()).toContain(urlSubString);
-            });
-        });
     }
 
     static async verifyAttributeContainsValue(targetElement: ElementFinder, attribute: string, expectedValue: string, stepLogger: StepLogger) {
         const actualValue = await PageHelper.getAttributeValue(targetElement, attribute);
         stepLogger.subVerification(`${actualValue} should contain to  ${expectedValue} value`);
-        await expect(actualValue).toContain(
+        expect(actualValue).toContain(
             expectedValue, ValidationsHelper.getValueContainsValidation(actualValue, expectedValue));
-    }
-
-    /**
-     * Verify that value is not empty
-     * @param {boolean} actualValue
-     * @param {StepLogger} stepLogger
-     * @returns {Promise<void>}
-     */
-    static async verifyBooleanValue(actualValue: boolean, expectedValue: boolean, stepLogger: StepLogger) {
-        stepLogger.subVerification(`${actualValue} should be ${expectedValue}`);
-        await expect(actualValue).toBe(expectedValue,
-            ValidationsHelper.getBooleamValidation(actualValue, expectedValue));
     }
 }
