+++ e2e/pom/pages/webapp/channel/channel-page.constants.ts
@@ -1,0 +1,36 @@
+export class ChannelPageConstants {
+
+    static get buttonLabel() {
+        return {
+            addToList: 'Add to list',
+            updateChannel: 'Update Channel',
+            createChannel: 'Create Channel',
+            copy: 'Copy',
+            continue: 'Continue',
+            addChannel: 'Add Channel',
+            addToNewChannel: 'Add to new Channel',
+            addToExistingChannel: 'Add to existing Channel',
+            add: 'Add',
+            relatedCompaniesAndTopics: 'Related companies and topics',
+            cancel: 'Cancel',
+            noThanks: 'No Thanks',
+            cancelUpdate: 'cancelupdate',
+            bulkRemove: 'Bulk Remove',
+            ok: 'Ok',
+        };
+    }
+
+    static get setUpChannelHeader() {
+        return {
+            verticals: 'Set Up Your Verticals Channel',
+            territories: 'Set Up Your Territories Channel',
+            productLines: 'Set Up Your Product Lines Channel',
+            singleAccount: 'Set Up Your Single Account Focus',
+            companies: 'Set Up Your Companies Channel',
+            competitors: 'Set Up Your Competitors Channel',
+            custom: 'Set Up Your Custom Channel',
+            accounts: 'Set Up Your Accounts Channel',
+            setupUpPage: 'Set Up Your'
+        };
+    }
+}
