+++ e2e/pom/pages/webapp/common/web-common-page.constants.ts
@@ -9,19 +9,6 @@ export class WebCommonPageConstants {
                 email: user.qaUser.email,
                 password: user.qaUser.password,
             },
-
-            topicsWorkbenchUser: {
-                username: user.topicsWorkbenchUser.username,
-                password: user.topicsWorkbenchUser.password,
-            },
-            qaUserUsedForRoleChange: {
-                email: user.qaUserUsedForRoleChange.email,
-                password: user.qaUserUsedForRoleChange.password,
-            },
-            legacyUser: {
-                email: user.legacyUser.email,
-                password: user.legacyUser.password,
-            },
         };
     }
 
@@ -29,10 +16,8 @@ export class WebCommonPageConstants {
         return {
             accounts: 'Accounts',
             singleAccount: 'Single Account Focus',
-            companiesToWatch: 'Companies to Watch',
             fiftyCharacters: 'This Channel contains more than fifty characters in its name',
             channelNameFifty: 'This Channel contains more than fifty characters i',
-            channelNameSuffix: 'This is a free online calculator which counts the number of characters',
         };
     }
