+++ e2e/pom/pages/webapp/login-page/login.po.ts
@@ -1,8 +1,8 @@
 import {LoginPageConstants} from './login-page.constants';
-import {HtmlHelper} from '../../../../components/misc-utils/html-helper';
-import {TextElements} from '../../../../components/html/html-component-types/text-elements';
-import {element, By} from 'protractor';
-import {TextboxElements} from '../../../../components/html/html-component-types/textbox-elements';
+import {TextboxElements} from '../../../../component/html/textbox-elements';
+import {TextElements} from '../../../../component/html/text-elements';
+import {HtmlHelper} from '../../../../component/misc-utils/html-helper';
+import {By, element} from 'protractor';
 
 export class LoginPage {
 
@@ -15,7 +15,7 @@ export class LoginPage {
     }
 
     static get signInButton() {
-        return TextElements.getItemById(HtmlHelper.attributeValues.signIn);
+        return element(By.id(HtmlHelper.attributeValues.signIn));
     }
 
     static get demoLoginLink() {
@@ -25,31 +25,6 @@ export class LoginPage {
     static get mainContainer() {
         return {
             loginPage: element(By.className('fr-signup-padding clearfix')),
-        };
-    }
-
-    static get topicsWorkbenchFormControls() {
-        const formIds = LoginPageConstants.topicWorkbenchLoginFormIds;
-        return {
-            username: element(By.css('input[name=userName]')),
-            password: element(By.css('input[name=password]')),
-            signIn: element(By.id(`${formIds.signIn}`)),
-        };
-    }
-    static get labels() {
-        return {
-            username: 'Username',
-            password: 'Password',
-            signIn: 'SignIn',
-        };
-    }
-
-    static get legacyFormControls() {
-        const formIds = LoginPageConstants.legacyLoginFormIds;
-        return {
-            username: element(By.id(formIds.email)),
-            password: element(By.id(formIds.password)),
-            signIn: element(By.id(formIds.signIn)),
         };
     }
 }
