+++ package.json
@@ -4,7 +4,20 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "postinstall": "webdriver-manager update --versions.chrome 2.41 --gecko=false",
+    "clean": "rm -rf auto-generated coverage dist e2e-report *.log",
+    "clean:all": "npm run clean -- node_modules",
+    "build": "tsc -p ./",
+    "lint": "tslint  --project tsconfig.json --config tslint.json",
+    "lint:fix": "npm run lint -- --fix",
+    "validate": "npm-run-all --parallel build lint",
+    "e2e": "protractor protractor.conf.js",
+    "e2e:report": "allure generate auto-generated/allure-results -o e2e-report --clean"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "npm run validate"
+    }
   },
   "author": "viet nguyen",
   "license": "MIT",
@@ -13,16 +26,25 @@
     "@angular/core": "^4.2.4",
     "@types/jasmine": "^3.4.4",
     "@types/node": "^12.11.1",
+    "allure-commandline": "^2.13.0",
     "codelyzer": "^5.1.2",
+    "core-js": "^3.3.2",
+    "husky": "^3.0.9",
+    "jasmine-allure-reporter": "^1.0.2",
     "jasmine-core": "^3.5.0",
     "jasminewd2": "^2.2.0",
     "log4js": "^5.2.2",
     "protractor": "^5.4.2",
+    "rxjs": "^6.5.3",
+    "rxjs-compat": "^6.5.3",
     "ts-node": "^8.4.1",
     "tslint": "^5.20.0",
+    "tslint-jasmine-rules": "^1.6.0",
+    "tslint-no-circular-imports": "^0.7.0",
     "typescript": "^3.6.4"
   },
   "devDependencies": {
-    "jasmine-spec-reporter": "^4.2.1"
+    "jasmine-spec-reporter": "^4.2.1",
+    "npm-run-all": "^4.1.5"
   }
 }
