+++ protractor.local.conf.js
@@ -1,0 +1,18 @@
+const defaultConfigSetup = require('./core/config-setup/default-config-setup');
+const reportersSetup = require('./core/config-setup/reporters-setup');
+
+exports.config = {
+    restartBrowserBetweenTests: defaultConfigSetup.restartBrowserBetweenTests,
+    SELENIUM_PROMISE_MANAGER: defaultConfigSetup.SELENIUM_PROMISE_MANAGER,
+    allScriptsTimeout: defaultConfigSetup.allScriptsTimeout,
+    suites: defaultConfigSetup.suites,
+    multiCapabilities: defaultConfigSetup.multiCapabilities,
+    params: defaultConfigSetup.params,
+    directConnect: true,
+    baseUrl: defaultConfigSetup.baseUrl,
+    framework: defaultConfigSetup.framework,
+    jasmineNodeOpts: defaultConfigSetup.jasmineNodeOpts,
+    onPrepare: function() {
+        reportersSetup.configureAllReporters();
+    }
+};
